requirements.txt
Django==1.10
argparse==1.2.1
dj-database-url==0.4.2
djangorestframework==3.6.2
gunicorn==19.7.1
pytz==2017.2
whitenoise==3.3.0
wsgiref==0.1.2
